не надо лишнего в выполненных домашках

+++ SuperSoft/client/src/components/Common/Tasks/DetailedAnswerSubtask.tsx
@@ -2,7 +2,7 @@
 import {Alert, Button, CardImg, CardText} from "reactstrap";
 import {ISubtaskProps} from "./ISubtaskProps";
 import {SubtaskViewModel} from "../../../Typings/viewModels/SubtaskViewModel";
-import {makeObservable, observable, toJS} from "mobx";
+import {makeObservable, observable} from "mobx";
 import {observer} from "mobx-react";
 import {UserRole} from "../../../Typings/enums/UserRole";
 import {UserSubtaskReadModel} from "../../../Typings/readModels/UserSubtaskReadModel";
@@ -47,7 +47,7 @@ export class DetailedAnswerSubtask extends Component<ISubtaskProps> {
     }
     
     renderControlButton() {
-        if(this.props.store.userStore.currentUser.role === UserRole.Admin) {
+        if(this.props.store.userStore.currentUser.role === UserRole.Admin && !this.props.reviewMode) {
             return(
                 <i style={{marginLeft: '94%', width: '2%'}}
                    onClick={() => this.deleteSubtask()}
@@ -76,7 +76,8 @@ export class DetailedAnswerSubtask extends Component<ISubtaskProps> {
     }
 
     renderInputAnswerArea() {
-        if(this.props.store.userStore.currentUser.role !== UserRole.Admin) {
+        let role = this.props.store.userStore.currentUser.role;
+        if(role !== UserRole.Admin) {
             return(
                 <div className="col-12">
                     <textarea
@@ -84,6 +85,14 @@ export class DetailedAnswerSubtask extends Component<ISubtaskProps> {
                         defaultValue={this.userAnswer.answer}
                         className="answerInput"
                         onChange={(e) => this.inputAnswer(e)}/>
+                </div>
+            );
+        } else if(role === UserRole.Admin) {
+            return(
+                <div className="col-12">
+                    <span>
+                        {this.userAnswer.answer}
+                    </span>
                 </div>
             );
         }
@@ -121,7 +130,6 @@ export class DetailedAnswerSubtask extends Component<ISubtaskProps> {
                 {this.renderImage(subtask)}
                 <CardText>
                     <div className="row justify-content-center">
-                        {this.renderControlButton()}
                         {this.renderInputAnswerArea()}
                     </div>
                     <div className="row justify-content-center">
